+++ src/main/java/com/felypeganzert/cacapalavras/repository/PalavraRepository.java
@@ -4,6 +4,6 @@ import com.felypeganzert.cacapalavras.entidades.Palavra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PalavraRepository extends JpaRepository<Palavra, Long>{
+public interface PalavraRepository extends JpaRepository<Palavra, Integer>{
     
 }
